Add authentication and configuration modules
- Introduced `auth.py` for handling access token retrieval from Auth0 using client credentials.
- Added `config.py` to manage environment configurations and check for the existence of the `.env` file.
- Created `main.py` as the entry point for the application, integrating user operations.
- Implemented `user_operations.py` for user management functions including blocking, deleting, and revoking grants.
- Added utility functions in `utils.py` for argument validation and user ID reading.
- Removed `delete.py` as its functionality has been refactored into the new structure.
